More organization and starting on getting information about a room from the EMS API
