+++ helpers/soap.go
@@ -7,48 +7,19 @@ import (
 	"net/http"
 )
 
+// SoapEnvelope wraps XML SOAP requests
 type SoapEnvelope struct {
 	XMLName struct{} `xml:"http://schemas.xmlsoap.org/soap/envelope/ Envelope"`
 	Body    SoapBody
 }
 
+// SoapBody is where The XML body goes
 type SoapBody struct {
 	XMLName  struct{} `xml:"Body"`
 	Contents []byte   `xml:",innerxml"`
 }
 
-type AllBuildingsRequest struct {
-	XMLName  struct{} `xml:"http://DEA.EMS.API.Web.Service/ GetBuildings"`
-	Username string   `xml:"UserName"`
-	Password string
-}
-
-type AllBuildingsResponse struct {
-	XMLName struct{} `xml:"http://DEA.EMS.API.Web.Service/ GetBuildingsResponse"`
-	Result  string   `xml:"GetBuildingsResult"`
-}
-
-type OneBuildingsRequest struct {
-	XMLName  struct{} `xml:"http://DEA.EMS.API.Web.Service/ GetBuildings"`
-	Username string   `xml:"UserName"`
-	Password string
-}
-
-type OneBuildingsResponse struct {
-	XMLName struct{} `xml:"http://DEA.EMS.API.Web.Service/ GetBuildingsResponse"`
-	Result  string   `xml:"GetBuildingsResult"`
-}
-
-type AllBuildings struct {
-	Buildings []Building `xml:"Data"`
-}
-
-type Building struct {
-	Building    string `xml:"BuildingCode"`
-	ID          int    `xml:"ID"`
-	Description string `xml:"Description"`
-}
-
+// SoapEncode encodes a request to be sent in a SOAP request
 func SoapEncode(contents interface{}) ([]byte, error) {
 	data, err := xml.MarshalIndent(contents, "    ", "  ")
 	if err != nil {
@@ -60,6 +31,7 @@ func SoapEncode(contents interface{}) ([]byte, error) {
 	return xml.MarshalIndent(&env, "", "  ")
 }
 
+// SoapDecode decodes a response returned by a SOAP request
 func SoapDecode(data []byte, contents interface{}) error {
 	env := SoapEnvelope{Body: SoapBody{}}
 	err := xml.Unmarshal(data, &env)
@@ -69,6 +41,7 @@ func SoapDecode(data []byte, contents interface{}) error {
 	return xml.Unmarshal(env.Body.Contents, contents)
 }
 
+// SoapRequest sends a SOAP request and returns the response
 func SoapRequest(url string, payload []byte) []byte {
 	resp, err := http.Post(url, "text/xml", bytes.NewBuffer(payload))
 	CheckErr(err)
